Move api call to its own file
