+++ .vscode/settings.json
@@ -1,20 +1,20 @@
 {
   "workbench.colorCustomizations": {
-    "activityBar.activeBackground": "#d24808",
-    "activityBar.activeBorder": "#08d74a",
-    "activityBar.background": "#d24808",
-    "activityBar.foreground": "#e7e7e7",
-    "activityBar.inactiveForeground": "#e7e7e799",
-    "activityBarBadge.background": "#08d74a",
-    "activityBarBadge.foreground": "#15202b",
-    "statusBar.background": "#a13706",
-    "statusBar.foreground": "#e7e7e7",
-    "statusBarItem.hoverBackground": "#d24808",
-    "titleBar.activeBackground": "#a13706",
-    "titleBar.activeForeground": "#e7e7e7",
-    "titleBar.inactiveBackground": "#a1370699",
-    "titleBar.inactiveForeground": "#e7e7e799"
+    "activityBar.activeBackground": "#65c89b",
+    "activityBar.activeBorder": "#945bc4",
+    "activityBar.background": "#65c89b",
+    "activityBar.foreground": "#15202b",
+    "activityBar.inactiveForeground": "#15202b99",
+    "activityBarBadge.background": "#945bc4",
+    "activityBarBadge.foreground": "#e7e7e7",
+    "statusBar.background": "#42b883",
+    "statusBar.foreground": "#15202b",
+    "statusBarItem.hoverBackground": "#359268",
+    "titleBar.activeBackground": "#42b883",
+    "titleBar.activeForeground": "#15202b",
+    "titleBar.inactiveBackground": "#42b88399",
+    "titleBar.inactiveForeground": "#15202b99"
   },
-  "peacock.color": "#a13706",
+  "peacock.color": "#42b883",
   "git.ignoreLimitWarning": true
 }
