+++ src/apiCalls.js
@@ -1,0 +1,5 @@
+export const getMovies = () => {
+  return fetch('https://rancid-tomatillos.herokuapp.com/api/v2/movies')
+    .then(response => response.json())
+}
+
